% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtable.r
\name{gtable_rbind}
\alias{gtable_rbind}
\alias{gtable_cbind}
\title{rbind gtables}
\usage{
gtable_rbind(..., size = "max", z = NULL)

gtable_cbind(..., size = "max", z = NULL)
}
\arguments{
\item{...}{gtables}

\item{size}{how should the widths be calculated?
\code{max} maximum of all widths
\code{min} minimum of all widths
\code{first} widths/heights of first gtable
\code{last} widths/heights of last gtable}

\item{z}{optional z level}
}
\description{
rbind gtables

cbind gtables
}
