% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gtable.r
\name{cbind.gtable}
\alias{cbind.gtable}
\title{cbind gtables}
\usage{
cbind.gtable(..., size = "max", z = NULL)
}
\arguments{
\item{...}{gtables}

\item{size}{how should the heights be calculated?
\code{max} maximum of all heights
\code{min} minimum of all heights
\code{first} heights of first gtable
\code{last} heights of last gtable}

\item{z}{optional z level}
}
\description{
cbind gtables
}

