% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtable.r
\name{combine}
\alias{combine}
\title{Combine gtables based on row/column names.
 @param ... gtables
 @param along dimension to align along, \code{1} = rows,
  \code{2} = cols. 
 @param join when x and y have different names, how should the difference be resolved? 
 \code{inner} keep names that appear in both, 
 \code{outer} keep names that appear in either, 
 \code{left} keep names from \code{x}, 
 and \code{right} keep names from \code{y}.
 @export}
\usage{
combine(..., along = 1L, join = "outer")
}
\description{
Combine gtables based on row/column names.
 @param ... gtables
 @param along dimension to align along, \code{1} = rows,
  \code{2} = cols. 
 @param join when x and y have different names, how should the difference be resolved? 
 \code{inner} keep names that appear in both, 
 \code{outer} keep names that appear in either, 
 \code{left} keep names from \code{x}, 
 and \code{right} keep names from \code{y}.
 @export
}

