\name{expand.arguments}
\alias{expand.arguments}
\title{expand a list of arguments to full length (and misc.}
\usage{expand.arguments(...)
}
\description{expand a list of arguments to full length (and misc. undocumented functions)}
\alias{updatelist}
\alias{textii}
\alias{rectii}
\alias{pchlist}
\value{...}
\arguments{\item{...}{arguments}
}

