\name{gridExtra-package}
\alias{gridExtra-package}
\alias{gridExtra-package}
\title{misc. high-level functions for Grid graphics}
\description{functions for Grid graphics}
\details{\tabular{ll}{
Package: \tab gridExtra\cr
Type: \tab Package\cr
Version: \tab 0.6.5\cr
Date: \tab 2009-08-17\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab false\cr
}

functions in Grid}
\alias{gridExtra}
\docType{package}
\author{baptiste Auguie \email{baptiste.auguie@gmail.com}}
\references{R Graphics by Paul Murrell, ggplot2 source code}
\keyword{packagelibrary}
\seealso{\code{\link{Grid}}}

