% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gtable.r
\name{gtable_align}
\alias{gtable_align}
\title{Align two gtables based on their row/col names.}
\usage{
gtable_align(x, y, along = 1L, join = "outer")
}
\arguments{
\item{x}{\code{\link{gtable}}}

\item{y}{\code{\link{gtable}}}

\item{along}{dimension to align along, \code{1} = rows, \code{2} =' cols.}

\item{join}{when x and y have different names, how should the d'ifference be resolved?
\code{inner} keep names that appear in both,
\code{outer} keep names that appear in either,
\code{left} keep names from \code{x},
and \code{right} keep names from \code{y}.}
}
\value{
a list with elements \code{x} and \code{y}
 corresponding to the input gtables with extra rows/columns so that they now align.
}
\description{
Align two gtables based on their row/col names.
}
\seealso{
\code{\link{gtable_join}} to return the two gtables
  combined into a single gtable.
}

