% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gtable.r
\name{gtable_join}
\alias{gtable_join}
\title{Join two gtables together based on row/column names.}
\usage{
gtable_join(x, y, along = 1L, join = "outer")
}
\arguments{
\item{x}{\code{\link{gtable}}}

\item{y}{\code{\link{gtable}}}

\item{along}{dimension to align along, \code{1} = rows,
\code{2} = cols. Join will occur perpendicular to this direction.}

\item{join}{when x and y have different names, how should the d'ifference be resolved?
\code{inner} keep names that appear in both,
\code{outer} keep names that appear in either,
\code{left} keep names from \code{x},
and \code{right} keep names from \code{y}.}
}
\description{
Join two gtables together based on row/column names.
}

