% Generated by roxygen2 (4.0.0): do not edit by hand
\name{gtable_table}
\alias{gtable_table}
\title{Graphical display of a textual table.}
\usage{
gtable_table(d, widths, heights, fg.par = list(col = "black"),
  bg.par = list(fill = "grey98", col = NA), padding = unit(c(4, 4), "mm"))
}
\arguments{
  \item{d}{data.frame or matrix}

  \item{widths}{optional vector to specify column widths}

  \item{heights}{optional vector to specify row heights}

  \item{fg.par}{control parameters for text grobs}

  \item{bg.par}{control parameters for rect grobs}

  \item{padding}{unit of length 2}
}
\description{
Graphical display of a textual table.
}
\examples{
\donttest{
d <- head(iris, 3)
core <- gtable_table(d,
                     fg.par = list(col=1:5, fontsize=c(10,12,15)),
                     bg.par = list(fill=1:2, alpha=0.5))
colhead <- gtable_table(t(colnames(d)))
rowhead <- gtable_table(c("", rownames(d)))
g <- rbind(colhead, core)
g <- cbind(rowhead, g)
grid.newpage()
grid.draw(g)
g2 <- gtable_table(t(colnames(d)), widths = unit(rep(1, ncol(d)), "null"))
grid.newpage()
grid.draw(g)
}
}

