% Generated by roxygen2 (4.0.0): do not edit by hand
\name{marrangeGrob}
\alias{marrangeGrob}
\alias{print.arrangelist}
\title{marrangeGrob}
\usage{
marrangeGrob(grobs, ncol, nrow, perpage, ..., top = quote(paste("page", g,
  "of", pages)))
}
\arguments{
  \item{...}{args to \link{arrangeGrob}}

  \item{top}{see \link{arrangeGrob}}
}
\value{
list of class arrangelist
}
\description{
Interface to arrangeGrob that can dispatch on multiple pages
}
\details{
If the layout specifies both nrow and ncol, the list of grobs can be split
in multiple pages. Interactive devices print open new windows, whilst non-interactive
devices such as pdf call grid.newpage() between the drawings.
}
\examples{
\dontrun{
require(ggplot2)
pl <- lapply(1:11, function(.x) qplot(1:10,rnorm(10), main=paste("plot",.x)))
ml <- marrangeGrob(pl, nrow=2, ncol=2, 4)
## interactive use; open new devices
ml
## non-interactive use, multipage pdf
ggsave("multipage.pdf", ml)
}
}
\author{
baptiste Auguie
}

