\name{polygon.regular}
\title{polygon.regular}
\usage{polygon.regular(sides=4, scale.area=TRUE, star=FALSE)}

\description{
  compute vertices of a polygon
}
\alias{polygon.regular}
\value{a data.frame}
\seealso{\code{star}}
\references{ngon from the 'maptree' package}
\arguments{
  \item{sides}{int}
  \item{scale.area}{logical}
  \item{star}{logical}
}
\examples{polygon.regular(5)}
