\name{polygon.star}
\title{polygon.star}
\usage{polygon.star(n=4, adj=pi/2, r=1)}

\description{
  compute vertices of a starred polygon
}
\alias{polygon.star}
\value{a data.frame}
\seealso{\code{polygon.regular}}
\arguments{
  \item{n}{int}
  \item{adj}{numeric angle}
  \item{r}{radius}
}
\examples{polygon.star(5)}
