\name{polygon1}
\title{polygon1}
\usage{polygon1(n=5, ang=0, ...)}

\description{
  compute vertices of a convex polygon
}
\alias{polygon1}
\alias{polygon2}
\value{a data.frame}
\seealso{\code{polygon.regular}}
\arguments{
  \item{n}{integer}
  \item{ang}{numeric angle}
  \item{...}{unused}
}
\examples{polygon2(5)}
