% Generated by roxygen2 (4.0.0): do not edit by hand
\name{polygon_regular}
\alias{polygon_regular}
\title{polygon_regular}
\usage{
polygon_regular(n = 5, ang = 0, x = 0, y = 0,
  rotateFromOrigin = FALSE)
}
\arguments{
  \item{n}{integer}

  \item{ang}{numeric angle}

  \item{rotateFromOrigin}{logical}
}
\value{
a data.frame
}
\description{
compute vertices of a convex polygon
}
\examples{
polygon2(5)
}

