% Generated by roxygen2 (4.0.0): do not edit by hand
\name{read.tiff}
\alias{as.raster.pixmapRGB}
\alias{read.tiff}
\title{read tiff files with pixmap}
\usage{
read.tiff(con, ...)
}
\arguments{
  \item{con}{connection}

  \item{...}{unused}
}
\value{
...
}
\description{
read tiff files with pixmap
}

