\name{rowMax.units}
\alias{rowMax.units}
\title{rowMax.units}
\usage{rowMax.units(u, nrow)
}
\description{calculates the max of a list of units arranged in a matrix}
\alias{rowMax.units,}
\alias{colMax.units}
\value{a vector of units}
\seealso{\code{unit.c}, \code{unit}}
\arguments{\item{u}{list of units}
\item{nrow}{nrow}
}

