% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtable.r
\name{str.gtable}
\alias{str.gtable}
\title{Prints summary information on gtable objects
 @param object a gtable
 @param ... unused
 @importFrom utils str
 @export}
\usage{
\method{str}{gtable}(object, ...)
}
\description{
Prints summary information on gtable objects
 @param object a gtable
 @param ... unused
 @importFrom utils str
 @export
}

