\name{theme.default}
\title{theme}
\usage{theme.default(...)}

\description{
  themes for table grob
}
\alias{theme.default}
\alias{theme.grey}
\alias{theme.list}
\alias{theme.blank}
\alias{theme.vertical}
\alias{theme.black}
\alias{theme.white}
\value{theme}
\arguments{
  \item{...}{optional params to overwrite the theme defaults}
}
