% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{virtualGrob}
\alias{drawDetails.virtual}
\alias{grobHeight.virtual}
\alias{grobWidth.virtual}
\alias{heightDetails.virtual}
\alias{is.zero}
\alias{virtualGrob}
\alias{widthDetails.virtual}
\title{virtualGrob}
\format{\preformatted{List of 3
 $ name: chr "NULL"
 $ gp  : NULL
 $ vp  : NULL
 - attr(*, "class")= chr [1:3] "virtual" "grob" "gDesc"
}}
\usage{
virtualGrob
}
\description{
zero grob borrowed from ggplot2
}
\seealso{
Other grob userlevel: \code{\link{arrangeTableGrobs}},
  \code{\link{drawDetails.table}},
  \code{\link{grid.table}},
  \code{\link{heightDetails.table}},
  \code{\link{makeTableGrobs}}, \code{\link{tableGrob}},
  \code{\link{widthDetails.table}};
  \code{\link{barbedGrob}}, \code{\link{grid.barbed}};
  \code{\link{borderGrob}}, \code{\link{grid.border}};
  \code{\link{ellipseGrob}}, \code{\link{grid.ellipse}},
  \code{\link{grid.ngon}}, \code{\link{ngonGrob}};
  \code{\link{grid.stext}}, \code{\link{stextGrob}};
  \code{\link{interleaven}}
}
\keyword{datasets}

