% Generated by roxygen2 (4.0.0): do not edit by hand
\name{arcTextGrob}
\alias{arcTextGrob}
\alias{grid.arcText}
\title{place text labels on a circle and display relations with arcs}
\usage{
arcTextGrob(x = unit(0.5, "npc"), y = unit(0.5, "npc"), labels = LETTERS,
  links = sample(seq_along(labels), 50, replace = TRUE),
  default.units = "npc", gp = gpar(), ...)
}
\arguments{
  \item{x}{x unit}

  \item{y}{y unit}

  \item{labels}{text labels}

  \item{links}{integer vector}

  \item{default.units}{default units}

  \item{gp}{gpar}

  \item{...}{additional params (unused)}
}
\value{
grob
}
\description{
place text labels on a circle and display relations with arcs
}
\examples{
library(grid)
set.seed(1234)
grid.newpage()
grid.arcText()
}
\seealso{
Other grob userlevel: \code{\link{arrangeTableGrobs}},
  \code{\link{drawDetails.table}},
  \code{\link{grid.table}},
  \code{\link{heightDetails.table}},
  \code{\link{makeTableGrobs}}, \code{\link{tableGrob}},
  \code{\link{widthDetails.table}};
  \code{\link{barbedGrob}}, \code{\link{grid.barbed}};
  \code{\link{borderGrob}}, \code{\link{grid.border}};
  \code{\link{colorstripGrob}},
  \code{\link{grid.colorstrip}};
  \code{\link{drawDetails.pattern}},
  \code{\link{grid.pattern}},
  \code{\link{heightDetails.pattern}},
  \code{\link{patternGrob}},
  \code{\link{segments.pattern}},
  \code{\link{widthDetails.pattern}};
  \code{\link{drawDetails.rpattern}},
  \code{\link{grid.rpattern}},
  \code{\link{heightDetails.rpattern}},
  \code{\link{rpatternGrob}}, \code{\link{tile.motif}},
  \code{\link{widthDetails.rpattern}};
  \code{\link{drawDetails.virtual}},
  \code{\link{grobHeight.virtual}},
  \code{\link{grobWidth.virtual}},
  \code{\link{heightDetails.virtual}},
  \code{\link{is.zero}}, \code{\link{virtualGrob}},
  \code{\link{widthDetails.virtual}};
  \code{\link{ebimageGrob}}; \code{\link{ellipseGrob}},
  \code{\link{grid.ellipse}}; \code{\link{grid.ngon}},
  \code{\link{ngonGrob}}; \code{\link{grid.stext}},
  \code{\link{stextGrob}}; \code{\link{interleaven}};
  \code{\link{pixmapGrob}}
}

