% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/arrangeGrob.r
\name{arrangeGrob}
\alias{arrangeGrob}
\alias{grid.arrange}
\title{arrangeGrob}
\usage{
arrangeGrob(..., grobs = list(...), layout_matrix, vp = NULL,
  name = "arrange", as.table = TRUE, respect = FALSE, nrow = NULL,
  ncol = NULL, widths = NULL, heights = NULL, top = NULL,
  bottom = NULL, left = NULL, right = NULL)
}
\arguments{
\item{...}{grobs, and valid arguments to grid.layout}

\item{grobs}{list of grobs}

\item{layout_matrix}{optional layout}

\item{vp}{viewport}

\item{as.table}{logical: bottom-left to top-right or top-left to bottom-right}

\item{top}{string, or grob (requires a well-defined height, see example)}

\item{bottom}{string, or grob (requires a well-defined height, see example)}

\item{left}{string, or grob (requires a well-defined width, see example)}

\item{right}{string, or grob (requires a well-defined width, see example)}
}
\value{
return a gTree wrapping a gtable
}
\description{
arrange grobs on a page
}
\examples{
\dontrun{
require(ggplot2)
plots = lapply(1:5, function(.x) qplot(1:10,rnorm(10), main=paste("plot",.x)))
require(gridExtra)
do.call(grid.arrange,  plots)
require(lattice)
grid.arrange(qplot(1:10), xyplot(1:10~1:10),
tableGrob(head(iris)), nrow=2, as.table=TRUE, main="test main",
left = rectGrob(width=unit(1,"line)),
sub=textGrob("test sub", gp=gpar(font=2)))
}
}

