% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtable.r
\name{cbind.gtable}
\alias{cbind.gtable}
\title{cbind gtables
 @param ... gtables
 @param size how should the heights be calculated?
 \code{max} maximum of all heights
 \code{min} minimum of all heights
 \code{first} heights of first gtable
 \code{last} heights of last gtable
 @param z optional z level}
\usage{
cbind.gtable(..., size = "max", z = NULL)
}
\description{
cbind gtables
 @param ... gtables
 @param size how should the heights be calculated?
 \code{max} maximum of all heights
 \code{min} minimum of all heights
 \code{first} heights of first gtable
 \code{last} heights of last gtable
 @param z optional z level
}

