% Generated by roxygen2 (4.0.0): do not edit by hand
\name{colorstripGrob}
\alias{colorstripGrob}
\alias{grid.colorstrip}
\title{colorstrip}
\usage{
colorstripGrob(fill = 1:3, colour = fill, draw = TRUE, raster = FALSE,
  direction = c("vertical", "horizontal"))
}
\arguments{
  \item{fill}{vector of colours}

  \item{colour}{vector of colours}

  \item{draw}{logical}

  \item{direction}{direction (horizontal or vertical)}

  \item{raster}{logical, use grid.raster (with
  interpolation)}
}
\value{
a grob
}
\description{
strip of colors
}
\examples{
colorstripGrob(c("#E41A1C", "#377EB8", "#4DAF4A"))
}
\seealso{
\code{grid.rect}

Other grob userlevel: \code{\link{arcTextGrob}},
  \code{\link{grid.arcText}};
  \code{\link{arrangeTableGrobs}},
  \code{\link{drawDetails.table}},
  \code{\link{grid.table}},
  \code{\link{heightDetails.table}},
  \code{\link{makeTableGrobs}}, \code{\link{tableGrob}},
  \code{\link{widthDetails.table}};
  \code{\link{barbedGrob}}, \code{\link{grid.barbed}};
  \code{\link{borderGrob}}, \code{\link{grid.border}};
  \code{\link{drawDetails.pattern}},
  \code{\link{grid.pattern}},
  \code{\link{heightDetails.pattern}},
  \code{\link{patternGrob}},
  \code{\link{segments.pattern}},
  \code{\link{widthDetails.pattern}};
  \code{\link{drawDetails.rpattern}},
  \code{\link{grid.rpattern}},
  \code{\link{heightDetails.rpattern}},
  \code{\link{rpatternGrob}}, \code{\link{tile.motif}},
  \code{\link{widthDetails.rpattern}};
  \code{\link{drawDetails.virtual}},
  \code{\link{grobHeight.virtual}},
  \code{\link{grobWidth.virtual}},
  \code{\link{heightDetails.virtual}},
  \code{\link{is.zero}}, \code{\link{virtualGrob}},
  \code{\link{widthDetails.virtual}};
  \code{\link{ebimageGrob}}; \code{\link{ellipseGrob}},
  \code{\link{grid.ellipse}}; \code{\link{grid.ngon}},
  \code{\link{ngonGrob}}; \code{\link{grid.stext}},
  \code{\link{stextGrob}}; \code{\link{interleaven}};
  \code{\link{pixmapGrob}}
}

