% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtable.r
\name{gtable_combine}
\alias{gtable_combine}
\alias{combine}
\alias{combine}
\title{Combine gtables based on row/column names.}
\usage{
gtable_combine(..., along = 1L, join = "outer")

combine(..., along = 1L, join = "outer")
}
\arguments{
\item{...}{gtables}

\item{along}{dimension to align along, \code{1} = rows,
\code{2} = cols.}

\item{join}{when x and y have different names, how should the difference be resolved? 
\code{inner} keep names that appear in both, 
\code{outer} keep names that appear in either, 
\code{left} keep names from \code{x}, 
and \code{right} keep names from \code{y}.}
}
\description{
Combine gtables based on row/column names.
}
