% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ellipseGrob}
\alias{ellipseGrob}
\alias{grid.ellipse}
\title{ellipseGrob}
\usage{
ellipseGrob(x, y, size = 1, angle = pi/2, ar = 1, sides = 50,
  gp = gpar(), default.units = "native", size.units = "mm")
}
\arguments{
  \item{x}{x unit}

  \item{y}{y unit}

  \item{size}{size}

  \item{angle}{angle}

  \item{ar}{aspec ratio}

  \item{sides}{default to 50}

  \item{gp}{gpar}

  \item{default.units}{default units}

  \item{size.units}{size units}
}
\value{
grob
}
\description{
an ellipse grob
}
\examples{
g = ellipseGrob(1:10/11,1:10/11,size=5,ar=1:5,angle=rnorm(10),
                def="npc", gp=gpar(fill=grey(1:10/11)))
grid.draw(g)
}
\seealso{
Other grob userlevel: \code{\link{arrangeTableGrobs}},
  \code{\link{drawDetails.table}},
  \code{\link{grid.table}},
  \code{\link{heightDetails.table}},
  \code{\link{makeTableGrobs}}, \code{\link{tableGrob}},
  \code{\link{widthDetails.table}};
  \code{\link{barbedGrob}}, \code{\link{grid.barbed}};
  \code{\link{borderGrob}}, \code{\link{grid.border}};
  \code{\link{drawDetails.virtual}},
  \code{\link{grobHeight.virtual}},
  \code{\link{grobWidth.virtual}},
  \code{\link{heightDetails.virtual}},
  \code{\link{is.zero}}, \code{\link{virtualGrob}},
  \code{\link{widthDetails.virtual}};
  \code{\link{ellipseGrob}}, \code{\link{grid.ellipse}},
  \code{\link{grid.ngon}}, \code{\link{ngonGrob}};
  \code{\link{grid.stext}}, \code{\link{stextGrob}};
  \code{\link{interleaven}}
}

