\name{expand.arguments}
\alias{expand.arguments}
\alias{pchlist}
\alias{rectii}
\alias{textii}
\alias{updatelist}
\title{expand.arguments}
\usage{
  expand.arguments(...)
}
\arguments{
  \item{...}{arguments}
}
\value{
  ...
}
\description{
  expand a list of arguments to full length (and misc.
  undocumented functions)
}

