\name{expand.arguments}
\title{expand.arguments}
\usage{expand.arguments(...)}

\description{
  expand a list of arguments to full length (and misc.
  undocumented functions)
}
\alias{expand.arguments}
\alias{updatelist}
\alias{textii}
\alias{rectii}
\alias{pchlist}
\value{...}
\arguments{
  \item{...}{arguments}
}
