% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gridExtra-package.r
\docType{package}
\name{gridExtra-package}
\alias{gridExtra}
\alias{gridExtra-package}
\title{Misc. high-level functions for grid graphics}
\description{
Misc. high-level functions for grid graphics
}
\author{
baptiste Auguie \email{baptiste.auguie@gmail.com}
}
\references{
R Graphics by Paul Murrell (Chapman & Hall/CRC, August 2005)
}
\seealso{
\code{\link{Grid}}
}
\keyword{packagelibrary}

