% Generated by roxygen2 (4.0.0): do not edit by hand

\docType{package}
\name{gridExtra-package}
\alias{gridExtra}
\alias{gridExtra-package}
\title{misc. high-level functions for Grid graphics}
\description{
  functions for Grid graphics
}
\author{
  baptiste Auguie \email{baptiste.auguie@gmail.com}
}
\references{
  R Graphics by Paul Murrell, ggplot2 source code
}
\seealso{
  \code{\link{Grid}}
}
\keyword{packagelibrary}

