% Generated by roxygen2 (4.0.0): do not edit by hand
\name{interleaven}
\alias{interleaven}
\title{interleaven}
\usage{
interleaven(x = letters[1:3], y = 1:12, n = 4)
}
\arguments{
  \item{x}{vector}

  \item{y}{vector}

  \item{n}{integer}
}
\value{
a vector
}
\description{
mix two vectors
}
\examples{
interleaven(replicate(3, rectGrob(), simplify=FALSE), replicate(12, virtualGrob, simplify=FALSE), 4)
}
\seealso{
\code{rep}, \code{ggplot2:interleave}

Other grob userlevel: \code{\link{arcTextGrob}},
  \code{\link{grid.arcText}};
  \code{\link{arrangeTableGrobs}},
  \code{\link{drawDetails.table}},
  \code{\link{grid.table}},
  \code{\link{heightDetails.table}},
  \code{\link{makeTableGrobs}}, \code{\link{tableGrob}},
  \code{\link{widthDetails.table}};
  \code{\link{barbedGrob}}, \code{\link{grid.barbed}};
  \code{\link{borderGrob}}, \code{\link{grid.border}};
  \code{\link{colorstripGrob}},
  \code{\link{grid.colorstrip}};
  \code{\link{drawDetails.pattern}},
  \code{\link{grid.pattern}},
  \code{\link{heightDetails.pattern}},
  \code{\link{patternGrob}},
  \code{\link{segments.pattern}},
  \code{\link{widthDetails.pattern}};
  \code{\link{drawDetails.rpattern}},
  \code{\link{grid.rpattern}},
  \code{\link{heightDetails.rpattern}},
  \code{\link{rpatternGrob}}, \code{\link{tile.motif}},
  \code{\link{widthDetails.rpattern}};
  \code{\link{drawDetails.virtual}},
  \code{\link{grobHeight.virtual}},
  \code{\link{grobWidth.virtual}},
  \code{\link{heightDetails.virtual}},
  \code{\link{is.zero}}, \code{\link{virtualGrob}},
  \code{\link{widthDetails.virtual}};
  \code{\link{ebimageGrob}}; \code{\link{ellipseGrob}},
  \code{\link{grid.ellipse}}; \code{\link{grid.ngon}},
  \code{\link{ngonGrob}}; \code{\link{grid.stext}},
  \code{\link{stextGrob}}; \code{\link{pixmapGrob}}
}

