% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ngonGrob.r
\name{polygon_regular}
\alias{ellipseGrob}
\alias{grid.ellipse}
\alias{grid.ngon}
\alias{ngonGrob}
\alias{polygon_regular}
\title{ngonGrob}
\usage{
polygon_regular(n = 5, phase = 0)

ngonGrob(x, y, n = 5, size = 5, phase = pi/2, angle = pi/4, rho = 1,
  gp = gpar(colour = "black", fill = NA, linejoin = "mitre"), ...,
  position.units = "npc", size.units = "mm")

grid.ngon(...)

ellipseGrob(x, y, size = 5, angle = pi/4, rho = 1, n = 50,
  gp = gpar(colour = "black", fill = NA, linejoin = "mitre"), ...,
  position.units = "npc", size.units = "mm")

grid.ellipse(...)
}
\arguments{
\item{n}{number of vertices}

\item{phase}{angle of first point relative to x axis}

\item{x}{x unit}

\item{y}{y unit}

\item{size}{radius of circumscribing circle}

\item{rho}{aspect ratio}

\item{gp}{gpar}

\item{...}{further parameters passed to polygonGrob}

\item{size.units}{grid units for the sizes}

\item{default.units}{default units for the positions}
}
\value{
grob
}
\description{
Regular polygon grob
}
\section{Functions}{
\itemize{
\item \code{polygon_regular}: 

\item \code{grid.ngon}: 

\item \code{ellipseGrob}: 

\item \code{grid.ellipse}: 
}}
\examples{
N <- 8
xy <- polygon_regular(N)*2

# draw multiple polygons
g <- ngonGrob(unit(xy[,1],"cm") + unit(0.5,"npc"),
              unit(xy[,2],"cm") + unit(0.5,"npc"),
              n = seq_len(N) + 2)

grid.newpage()
grid.draw(g)

# rotated and stretched
g2 <- ngonGrob(unit(xy[,1],"cm") + unit(0.5,"npc"),
              unit(xy[,2],"cm") + unit(0.5,"npc"),
              n = seq_len(N) + 2,
              phase = pi/2, angle=pi/(seq_len(N)+2),
              size=1:N+5)

grid.newpage()
grid.draw(g2)

# ellipse
g3 <- ellipseGrob(unit(xy[,1],"cm") + unit(0.5,"npc"),
                  unit(xy[,2],"cm") + unit(0.5,"npc"),
                  angle=seq_len(N)* pi/N,
                  size=1:N+5, rho=1:N)

grid.newpage()
grid.draw(g3)
}

