% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ngonGrob}
\alias{ellipseGrob}
\alias{grid.ellipse}
\alias{grid.ngon}
\alias{ngonGrob}
\title{ngonGrob}
\usage{
ngonGrob(x, y, sides = 5, size = 1, angle = pi/2, ar = 1,
  gp = gpar(colour = "grey50", fill = "grey90", alpha = 0.5, linejoin =
  "mitre"), units.def = "native", units.size = "mm")
}
\arguments{
  \item{x}{x unit}

  \item{y}{y unit}

  \item{sides}{sides}

  \item{size}{size}

  \item{angle}{angle}

  \item{ar}{ar}

  \item{units.def}{default units for the positions}

  \item{units.size}{units for the polygon sizes}

  \item{gp}{gpar}
}
\value{
grob
}
\description{
Regular polygon grob
}
\examples{
pushViewport(dataViewport(0:1, 0:1, width=unit(2, "cm"),
                          height=unit(2, "cm")))

xy <- polygon_regular(6)
grid.ngon(0.5, 0.5, 6, 10, units.size="mm")
for(ii in 1:NROW(xy)){
 grid.ngon(xy[ii, 1]+0.5, xy[ii, 2]+0.5, 6, 10, units.size="mm")
}
upViewport()
}
\seealso{
Other grob userlevel: \code{\link{barbedGrob}},
  \code{\link{grid.barbed}}; \code{\link{borderGrob}},
  \code{\link{grid.border}}; \code{\link{grid.stext}},
  \code{\link{stextGrob}}
}

