% Generated by roxygen2 (4.0.0): do not edit by hand
\name{pixmapGrob}
\alias{pixmapGrob}
\title{create a grob from pixmap object}
\usage{
pixmapGrob(pic, x = 0.5, y = 0.5, scale = 1, raster = FALSE,
  angle = 0, vp = NULL, ...)
}
\arguments{
  \item{pic}{pixmap object}

  \item{x}{x unit}

  \item{y}{y unit}

  \item{scale}{scale}

  \item{raster}{raster}

  \item{angle}{angle}

  \item{vp}{viewport}

  \item{...}{optional grob parameters, passed to imageGrob
  or rasterGrob}
}
\value{
a gTree of class 'pixmap', with natural width and height in points
}
\description{
create a grob from pixmap object
}
\details{
Very primitive function,  using RGraphics' imageGrob or rasterGrob (R>=2.11)
}
\examples{
library(pixmap)
library(RGraphics)
x <- read.pnm(system.file("pictures/logo.ppm", package="pixmap")[1])
g1 <- pixmapGrob(x)
dev.new(width=g1$width/72, height=g1$height/72)
grid.draw(g1)
}
\seealso{
Other grob userlevel: \code{\link{arcTextGrob}},
  \code{\link{grid.arcText}};
  \code{\link{arrangeTableGrobs}},
  \code{\link{drawDetails.table}},
  \code{\link{grid.table}},
  \code{\link{heightDetails.table}},
  \code{\link{makeTableGrobs}}, \code{\link{tableGrob}},
  \code{\link{widthDetails.table}};
  \code{\link{barbedGrob}}, \code{\link{grid.barbed}};
  \code{\link{borderGrob}}, \code{\link{grid.border}};
  \code{\link{colorstripGrob}},
  \code{\link{grid.colorstrip}};
  \code{\link{drawDetails.pattern}},
  \code{\link{grid.pattern}},
  \code{\link{heightDetails.pattern}},
  \code{\link{patternGrob}},
  \code{\link{segments.pattern}},
  \code{\link{widthDetails.pattern}};
  \code{\link{drawDetails.rpattern}},
  \code{\link{grid.rpattern}},
  \code{\link{heightDetails.rpattern}},
  \code{\link{rpatternGrob}}, \code{\link{tile.motif}},
  \code{\link{widthDetails.rpattern}};
  \code{\link{drawDetails.virtual}},
  \code{\link{grobHeight.virtual}},
  \code{\link{grobWidth.virtual}},
  \code{\link{heightDetails.virtual}},
  \code{\link{is.zero}}, \code{\link{virtualGrob}},
  \code{\link{widthDetails.virtual}};
  \code{\link{ebimageGrob}}; \code{\link{ellipseGrob}},
  \code{\link{grid.ellipse}}; \code{\link{grid.ngon}},
  \code{\link{ngonGrob}}; \code{\link{grid.stext}},
  \code{\link{stextGrob}}; \code{\link{interleaven}}
}

