% Generated by roxygen2 (4.0.0): do not edit by hand
\name{polygon.star}
\alias{polygon.star}
\title{polygon.star}
\usage{
polygon.star(n = 4, adj = pi/2, r = 1)
}
\arguments{
  \item{n}{int}

  \item{adj}{numeric angle}

  \item{r}{radius}
}
\value{
a data.frame
}
\description{
compute vertices of a starred polygon
}
\examples{
polygon.star(5)
}

