% Generated by roxygen2 (4.0.0): do not edit by hand
\name{polygon1}
\alias{polygon1}
\alias{polygon2}
\title{polygon1}
\usage{
polygon1(n = 5, ang = 0, ...)
}
\arguments{
  \item{n}{integer}

  \item{ang}{numeric angle}

  \item{...}{unused}
}
\value{
a data.frame
}
\description{
compute vertices of a convex polygon
}
\examples{
polygon2(5)
}

