% Generated by roxygen2 (4.0.0): do not edit by hand

\name{rowMax.units}
\alias{colMax.units}
\alias{rowMax.units}
\title{rowMax.units}
\usage{
  rowMax.units(u, nrow)
}
\arguments{
  \item{u}{list of units}

  \item{nrow}{nrow}
}
\value{
  a vector of units
}
\description{
  calculates the max of a list of units arranged in a
  matrix
}
\seealso{
  \code{unit.c}, \code{unit}
}

