% Generated by roxygen2 (4.0.0): do not edit by hand
\name{stextGrob}
\alias{grid.stext}
\alias{stextGrob}
\title{stextGrob}
\usage{
stextGrob(label, r = 0.1, x = unit(0.5, "npc"), y = unit(0.5, "npc"),
  just = "centre", hjust = NULL, vjust = NULL, rot = 0,
  check.overlap = FALSE, default.units = "npc", name = NULL,
  gp = gpar(), vp = NULL)
}
\arguments{
  \item{label}{see textGrob}

  \item{r}{blur radius}

  \item{x}{see textGrob}

  \item{y}{see textGrob}

  \item{just}{see textGrob}

  \item{hjust}{see textGrob}

  \item{vjust}{see textGrob}

  \item{rot}{see textGrob}

  \item{check.overlap}{see textGrob}

  \item{default.units}{see textGrob}

  \item{name}{see textGrob}

  \item{gp}{see textGrob}

  \item{vp}{see textGrob}
}
\value{
gTree
}
\description{
shadow text
}
\details{
adds a blurred white version of a label below the text
}
\examples{
grid.rect(gp=gpar(fill="grey"))
grid.stext("test")
}
\author{
Baptiste Auguie
}
\seealso{
Other grob userlevel: \code{\link{arcTextGrob}},
  \code{\link{grid.arcText}};
  \code{\link{arrangeTableGrobs}},
  \code{\link{drawDetails.table}},
  \code{\link{grid.table}},
  \code{\link{heightDetails.table}},
  \code{\link{makeTableGrobs}}, \code{\link{tableGrob}},
  \code{\link{widthDetails.table}};
  \code{\link{barbedGrob}}, \code{\link{grid.barbed}};
  \code{\link{borderGrob}}, \code{\link{grid.border}};
  \code{\link{colorstripGrob}},
  \code{\link{grid.colorstrip}};
  \code{\link{drawDetails.pattern}},
  \code{\link{grid.pattern}},
  \code{\link{heightDetails.pattern}},
  \code{\link{patternGrob}},
  \code{\link{segments.pattern}},
  \code{\link{widthDetails.pattern}};
  \code{\link{drawDetails.rpattern}},
  \code{\link{grid.rpattern}},
  \code{\link{heightDetails.rpattern}},
  \code{\link{rpatternGrob}}, \code{\link{tile.motif}},
  \code{\link{widthDetails.rpattern}};
  \code{\link{drawDetails.virtual}},
  \code{\link{grobHeight.virtual}},
  \code{\link{grobWidth.virtual}},
  \code{\link{heightDetails.virtual}},
  \code{\link{is.zero}}, \code{\link{virtualGrob}},
  \code{\link{widthDetails.virtual}};
  \code{\link{ebimageGrob}}; \code{\link{ellipseGrob}},
  \code{\link{grid.ellipse}}; \code{\link{grid.ngon}},
  \code{\link{ngonGrob}}; \code{\link{interleaven}};
  \code{\link{pixmapGrob}}
}

