% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtable.r
\name{str.gtable}
\alias{str.gtable}
\title{Prints summary information on gtable objects}
\usage{
\method{str}{gtable}(object, ...)
}
\arguments{
\item{object}{a gtable}

\item{...}{unused}
}
\description{
Prints summary information on gtable objects
}

