% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tableGrob.r
\name{tableGrob}
\alias{tableGrob}
\title{Graphical display of a textual table.}
\usage{
tableGrob(d, rows = rownames(d), cols = colnames(d),
  theme = ttheme_default(), ...)
}
\arguments{
\item{d}{data.frame or matrix}

\item{widths}{optional vector to specify column widths}

\item{heights}{optional vector to specify row heights}

\item{fg.par}{control parameters for text grobs}

\item{bg.par}{control parameters for rect grobs}

\item{padding}{unit of length 2}
}
\description{
Graphical display of a textual table.
}
\examples{
\donttest{
d <- head(iris, 3)
g <- tableGrob(d)
grid.newpage()
grid.draw(g)
}
}

