\name{theme.default}
\alias{theme.black}
\alias{theme.blank}
\alias{theme.default}
\alias{theme.grey}
\alias{theme.list}
\alias{theme.vertical}
\alias{theme.white}
\title{theme}
\usage{
  theme.default(...)
}
\arguments{
  \item{...}{optional params to overwrite the theme
  defaults}
}
\value{
  theme
}
\description{
  themes for table grob
}

