% Generated by roxygen2 (4.0.0): do not edit by hand

\docType{data}
\name{virtualGrob}
\alias{drawDetails.virtual}
\alias{grobHeight.virtual}
\alias{grobWidth.virtual}
\alias{heightDetails.virtual}
\alias{is.zero}
\alias{virtualGrob}
\alias{widthDetails.virtual}
\title{virtualGrob}
\format{List of 3
 $ name: chr "NULL"
 $ gp  : NULL
 $ vp  : NULL
 - attr(*, "class")= chr [1:3] "virtual" "grob" "gDesc"}
\usage{
  virtualGrob
}
\description{
  zero grob borrowed from ggplot2
}
\seealso{
  Other grob userlevel: \code{\link{arcTextGrob}},
  \code{\link{barbedGrob}}, \code{\link{borderGrob}},
  \code{\link{colorstripGrob}}, \code{\link{ebimageGrob}},
  \code{\link{ellipseGrob}}, \code{\link{interleaven}},
  \code{\link{ngonGrob}}, \code{\link{patternGrob}},
  \code{\link{pixmapGrob}}, \code{\link{rpatternGrob}},
  \code{\link{stextGrob}}, \code{\link{tableGrob}}
}
\keyword{datasets}

